refactor: introduce bundle geometry builder
