+++ bpy_speckle/converter/from_bundle/_baking/geometry/__init__.py
@@ -1,1 +1,5 @@
-"""Geometry construction internals."""
+"""The private geometry construction interface."""
+
+from .builder import GeometryBuilder
+
+__all__ = ["GeometryBuilder"]
